requirements.txt
altgraph==0.17.3
black==23.9.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
pefile==2023.2.7
Pillow==10.0.1
pillow-heif==0.13.0
platformdirs==3.2.0
Pygments==2.14.0
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
pywin32-ctypes==0.2.2
rich==12.6.0
shellingham==1.5.0.post1
typer==0.9.0
typing_extensions==4.8.0
